Move ec2connect to a separate library
