+++ lib/ec2connect/connect.go
@@ -1,4 +1,4 @@
-package lib
+package ec2connect
 
 import (
 	"fmt"
